Rename if_re_bsd.h and rtlhw.cpp to match BSD driver
To make it easier to relate the BSD driver sources to the Windows driver
sources, restore the original BSD driver filenames:

- `rtlhw.cpp` -> `if_re.c`
- `if_re_bsd.h` -> `if_rereg.h`

Also, compile the BSD sources using C rules instead of C++ rules, which
hopefully make it easier to keep the BSD and Windows code aligned.

Compiling as C code requires a few fixes:

- Functions shared between C and C++ code need to be compiled as `extern
  "C"`.
- To use `bool` in C code, you have to include the `stdbool.h` header.
- During the build, the compiler can't use the same .pch for both C and
  C++ code, so only "precompile" the header for the C++ code (compile
  the C code without precompiled header acceleration).
