+++ if_re.vcxproj
@@ -210,7 +210,7 @@
     <ClInclude Include="ethernet.h" />
     <ClInclude Include="forward.h" />
     <ClInclude Include="if.h" />
-    <ClInclude Include="if_re_bsd.h" />
+    <ClInclude Include="if_rereg.h" />
     <ClInclude Include="interrupt.h" />
     <ClInclude Include="link.h" />
     <ClInclude Include="mbuf.h" />
@@ -238,7 +238,12 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="rtlhw.cpp" />
+    <ClCompile Include="if_re.c">
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|ARM64'">NotUsing</PrecompiledHeader>
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|ARM64'">NotUsing</PrecompiledHeader>
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
+      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
+    </ClCompile>
     <ClCompile Include="rxqueue.cpp" />
     <ClCompile Include="txqueue.cpp" />
   </ItemGroup>
