+++ if_re.vcxproj.Filters
@@ -56,7 +56,7 @@
     <ClCompile Include="txqueue.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="rtlhw.cpp">
+    <ClCompile Include="if_re.c">
       <Filter>Source Files\BSD</Filter>
     </ClCompile>
     <ClCompile Include="pciconfig.cpp">
@@ -72,6 +72,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="mcast.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="precomp.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -103,7 +106,7 @@
     <ClInclude Include="bsd.h">
       <Filter>Header Files\BSD</Filter>
     </ClInclude>
-    <ClInclude Include="if_re_bsd.h">
+    <ClInclude Include="if_rereg.h">
       <Filter>Header Files\BSD</Filter>
     </ClInclude>
     <ClInclude Include="mii.h">
